REST endpoint added
